Use router for `rooms` paths
