+++ routes/all.ts
@@ -1,0 +1,1 @@
+export {default as rooms}	from "./rooms";
